move SNS rule tests to new file & add test for deleting multiple SNS rules with same topic
