+++ packages/api/tests/models/rules/test-rules-model.js
@@ -498,236 +498,6 @@ test('Creating a rule with a queueName parameter', async (t) => {
   t.is(payload.queueName, ruleItem.queueName);
 });
 
-test('creating a disabled SNS rule creates no event source mapping', async (t) => {
-  const snsTopicArn = randomString();
-  const item = fakeRuleFactoryV2({
-    workflow,
-    rule: {
-      type: 'sns',
-      value: snsTopicArn
-    },
-    state: 'DISABLED'
-  });
-
-  const rule = await rulesModel.create(item);
-
-  t.is(rule.state, 'DISABLED');
-  t.is(rule.rule.value, snsTopicArn);
-  t.falsy(rule.rule.arn);
-});
-
-test.serial('disabling an SNS rule removes the event source mapping', async (t) => {
-  const snsTopicArn = randomString();
-  const snsStub = sinon.stub(awsServices, 'sns')
-    .returns({
-      listSubscriptionsByTopic: () => ({
-        promise: () => Promise.resolve({
-          Subscriptions: [{
-            Endpoint: process.env.messageConsumer,
-            SubscriptionArn: snsTopicArn
-          }]
-        })
-      }),
-      unsubscribe: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-  const lambdaStub = sinon.stub(awsServices, 'lambda')
-    .returns({
-      addPermission: () => ({
-        promise: () => Promise.resolve()
-      }),
-      removePermission: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-
-  const item = fakeRuleFactoryV2({
-    workflow,
-    rule: {
-      type: 'sns',
-      value: snsTopicArn
-    },
-    state: 'ENABLED'
-  });
-
-  const rule = await rulesModel.create(item);
-
-  t.is(rule.rule.value, snsTopicArn);
-  t.truthy(rule.rule.arn);
-  t.is(rule.state, 'ENABLED');
-
-  const updates = { name: rule.name, state: 'DISABLED' };
-  const updatedRule = await rulesModel.update(rule, updates);
-
-  t.is(updatedRule.name, rule.name);
-  t.is(updatedRule.state, 'DISABLED');
-  t.is(updatedRule.rule.type, rule.rule.type);
-  t.is(updatedRule.rule.value, rule.rule.value);
-  t.falsy(updatedRule.rule.arn);
-
-  await rulesModel.delete(rule);
-  snsStub.restore();
-  lambdaStub.restore();
-});
-
-test.serial('enabling a disabled SNS rule and passing rule.arn throws specific error', async (t) => {
-  const snsTopicArn = randomString();
-  const snsStub = sinon.stub(awsServices, 'sns')
-    .returns({
-      listSubscriptionsByTopic: () => ({
-        promise: () => Promise.resolve({
-          Subscriptions: [{
-            Endpoint: process.env.messageConsumer,
-            SubscriptionArn: snsTopicArn
-          }]
-        })
-      }),
-      unsubscribe: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-
-  const item = fakeRuleFactoryV2({
-    workflow,
-    rule: {
-      type: 'sns',
-      value: snsTopicArn
-    },
-    state: 'DISABLED'
-  });
-
-  const rule = await rulesModel.create(item);
-
-  t.is(rule.rule.value, snsTopicArn);
-  t.falsy(rule.rule.arn);
-  t.is(rule.state, 'DISABLED');
-
-  const updates = {
-    name: rule.name,
-    state: 'ENABLED',
-    rule: {
-      ...rule.rule,
-      arn: 'test-value'
-    }
-  };
-  try {
-    // Should fail because a disabled rule should not have an ARN
-    // when being updated
-    await t.throwsAsync(rulesModel.update(rule, updates), null,
-      'Including rule.arn is not allowed when enabling a disabled rule');
-  } finally {
-    await rulesModel.delete(rule);
-    snsStub.restore();
-  }
-});
-
-test.serial('updating an SNS rule updates the event source mapping', async (t) => {
-  const snsTopicArn = randomString();
-  const newSnsTopicArn = randomString();
-
-  const snsStub = sinon.stub(awsServices, 'sns')
-    .returns({
-      listSubscriptionsByTopic: () => ({
-        promise: () => Promise.resolve({
-          Subscriptions: [{
-            Endpoint: process.env.messageConsumer,
-            SubscriptionArn: randomString()
-          }]
-        })
-      }),
-      unsubscribe: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-  const lambdaStub = sinon.stub(awsServices, 'lambda')
-    .returns({
-      addPermission: () => ({
-        promise: () => Promise.resolve()
-      }),
-      removePermission: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-
-  const item = fakeRuleFactoryV2({
-    workflow,
-    rule: {
-      type: 'sns',
-      value: snsTopicArn
-    },
-    state: 'ENABLED'
-  });
-
-  const rule = await rulesModel.create(item);
-
-  t.is(rule.rule.value, snsTopicArn);
-
-  const updates = { name: rule.name, rule: { value: newSnsTopicArn } };
-  const updatedRule = await rulesModel.update(rule, updates);
-
-  t.is(updatedRule.name, rule.name);
-  t.is(updatedRule.type, rule.type);
-  t.is(updatedRule.rule.value, newSnsTopicArn);
-  t.not(updatedRule.rule.arn, rule.rule.arn);
-
-  await rulesModel.delete(rule);
-  snsStub.restore();
-  lambdaStub.restore();
-});
-
-test.serial('deleting an SNS rule updates the event source mapping', async (t) => {
-  const snsTopicArn = randomString();
-
-  const snsStub = sinon.stub(awsServices, 'sns')
-    .returns({
-      listSubscriptionsByTopic: () => ({
-        promise: () => Promise.resolve({
-          Subscriptions: [{
-            Endpoint: process.env.messageConsumer,
-            SubscriptionArn: randomString()
-          }]
-        })
-      }),
-      unsubscribe: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-  const lambdaStub = sinon.stub(awsServices, 'lambda')
-    .returns({
-      addPermission: () => ({
-        promise: () => Promise.resolve()
-      }),
-      removePermission: () => ({
-        promise: () => Promise.resolve()
-      })
-    });
-  const unsubscribeSpy = sinon.spy(awsServices.sns(), 'unsubscribe');
-
-  const item = fakeRuleFactoryV2({
-    workflow,
-    rule: {
-      type: 'sns',
-      value: snsTopicArn
-    },
-    state: 'ENABLED'
-  });
-
-  const rule = await rulesModel.create(item);
-
-  t.is(rule.rule.value, snsTopicArn);
-
-  await rulesModel.delete(rule);
-
-  t.true(unsubscribeSpy.called);
-  t.true(unsubscribeSpy.calledWith({
-    SubscriptionArn: rule.rule.arn
-  }));
-
-  snsStub.restore();
-  lambdaStub.restore();
-});
-
 test('updates rule meta object', async (t) => {
   const { onetimeRule } = t.context;
